requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
